chore(trace_utils): move implementation details to internal
