+++ ddtrace/contrib/internal/mysqldb/patch.py
@@ -9,7 +9,7 @@ from ddtrace.appsec._iast.constants import VULN_SQL_INJECTION
 from ddtrace.constants import SPAN_KIND
 from ddtrace.constants import SPAN_MEASURED_KEY
 from ddtrace.contrib.dbapi import TracedConnection
-from ddtrace.contrib.trace_utils import _convert_to_string
+from ddtrace.contrib.internal.trace_utils import _convert_to_string
 from ddtrace.contrib.trace_utils import ext_service
 from ddtrace.ext import SpanKind
 from ddtrace.ext import SpanTypes
