+++ ddtrace/contrib/internal/pymysql/patch.py
@@ -5,7 +5,7 @@ import wrapt
 
 from ddtrace import config
 from ddtrace.contrib.dbapi import TracedConnection
-from ddtrace.contrib.trace_utils import _convert_to_string
+from ddtrace.contrib.internal.trace_utils import _convert_to_string
 from ddtrace.ext import db
 from ddtrace.ext import net
 from ddtrace.internal.schema import schematize_database_operation
