+++ ddtrace/contrib/internal/redis/asyncio_patch.py
@@ -2,7 +2,7 @@ from ddtrace import config
 from ddtrace._trace.utils_redis import _instrument_redis_cmd
 from ddtrace._trace.utils_redis import _instrument_redis_execute_async_cluster_pipeline
 from ddtrace._trace.utils_redis import _instrument_redis_execute_pipeline
-from ddtrace.contrib.redis_utils import _run_redis_command_async
+from ddtrace.contrib.internal.redis_utils import _run_redis_command_async
 from ddtrace.internal.utils.formats import stringify_cache_args
 from ddtrace.trace import Pin
